+++ Sources/SuperwallKit/Paywall/Presentation/Internal/Presentation Request/PresentationRequest.swift
@@ -99,6 +99,7 @@ struct PresentationRequest {
 
   struct Flags {
     var isDebuggerLaunched: Bool
+    // TODO: Check if we can just use entitlements  (as that contains didSetActiveEntitlements)
     var didSetActiveEntitlements: AnyPublisher<Bool, Never>
     var entitlements: EntitlementsInfo
     var isPaywallPresented: Bool