+++ Tests/SuperwallKitTests/Config/ConfigManagerTests.swift
@@ -29,7 +29,8 @@ final class ConfigManagerTests: XCTestCase {
       storage: storage,
       network: network,
       paywallManager: dependencyContainer.paywallManager,
-      deviceHelper: dependencyContainer.deviceHelper,
+      deviceHelper: dependencyContainer.deviceHelper, 
+      entitlementsInfo: dependencyContainer.entitlementsInfo,
       factory: dependencyContainer
     )
     configManager.confirmAssignment(assignment)
@@ -56,6 +57,7 @@ final class ConfigManagerTests: XCTestCase {
       network: network,
       paywallManager: dependencyContainer.paywallManager,
       deviceHelper: dependencyContainer.deviceHelper,
+      entitlementsInfo: dependencyContainer.entitlementsInfo,
       factory: dependencyContainer
     )
 
@@ -82,7 +84,8 @@ final class ConfigManagerTests: XCTestCase {
       storage: storage,
       network: network,
       paywallManager: dependencyContainer.paywallManager,
-      deviceHelper: dependencyContainer.deviceHelper,
+      deviceHelper: dependencyContainer.deviceHelper, 
+      entitlementsInfo: dependencyContainer.entitlementsInfo,
       factory: dependencyContainer
     )
     configManager.configState.send(.retrieved(.stub()
@@ -105,6 +108,7 @@ final class ConfigManagerTests: XCTestCase {
       network: network,
       paywallManager: dependencyContainer.paywallManager,
       deviceHelper: dependencyContainer.deviceHelper,
+      entitlementsInfo: dependencyContainer.entitlementsInfo,
       factory: dependencyContainer
     )