+++ Tests/SuperwallKitTests/Paywall/Request/PaywallLogicTests.swift
@@ -52,8 +52,14 @@ class PaywallLogicTests: XCTestCase {
     let id = "myid"
     let locale = "en_US"
     let event: EventData = .stub()
-    let product1 = StoreProduct(sk1Product: MockSkProduct(productIdentifier: "abc"))
-    let product2 = StoreProduct(sk1Product: MockSkProduct(productIdentifier: "def"))
+    let product1 = StoreProduct(
+      sk1Product: MockSkProduct(productIdentifier: "abc"),
+      entitlements: []
+    )
+    let product2 = StoreProduct(
+      sk1Product: MockSkProduct(productIdentifier: "def"),
+      entitlements: []
+    )
     let ids = ["abc", "def"].joined()
 
     // When
@@ -106,7 +112,10 @@ class PaywallLogicTests: XCTestCase {
     let eventName = "MyEvent"
     let event: EventData = .stub()
       .setting(\.name, to: eventName)
-    let product1 = StoreProduct(sk1Product: MockSkProduct(productIdentifier: "abc"))
+    let product1 = StoreProduct(
+      sk1Product: MockSkProduct(productIdentifier: "abc"),
+      entitlements: []
+    )
     let ids = "abc"
 
     // When
@@ -144,7 +153,8 @@ class PaywallLogicTests: XCTestCase {
     let productId = "id1"
     let products = [Product(
       name: "primary",
-      type: .appStore(.init(id: productId))
+      type: .appStore(.init(id: productId)),
+      entitlements: []
     )]
 
     let skProductId = "id2"
@@ -152,8 +162,8 @@ class PaywallLogicTests: XCTestCase {
       productIdentifier: skProductId,
       price: 1.99
     )
-    let productsById = [skProductId: StoreProduct(sk1Product: skProduct)]
-    
+    let productsById = [skProductId: StoreProduct(sk1Product: skProduct, entitlements: [])]
+
     let response = await PaywallLogic.getVariablesAndFreeTrial(
       productItems: products,
       productsById: productsById,
@@ -177,7 +187,8 @@ class PaywallLogicTests: XCTestCase {
     let productId = "id1"
     let products = [Product(
       name: "secondary",
-      type: .appStore(.init(id: productId))
+      type: .appStore(.init(id: productId)),
+      entitlements: []
     )]
 
     let product = StoreProduct(
@@ -185,7 +196,8 @@ class PaywallLogicTests: XCTestCase {
         MockSkProduct(
           productIdentifier: productId,
           price: 1.99
-        )
+        ),
+      entitlements: []
     )
     let productsById = [productId: product]
 
@@ -212,7 +224,7 @@ class PaywallLogicTests: XCTestCase {
     let productId = "id1"
     let products = [Product(
       name: "primary",
-      type: .appStore(.init(id: productId))
+      type: .appStore(.init(id: productId)), entitlements: []
     )]
     let mockIntroPeriod = MockIntroductoryPeriod(
       testSubscriptionPeriod: MockSubscriptionPeriod()
@@ -224,7 +236,8 @@ class PaywallLogicTests: XCTestCase {
           productIdentifier: productId,
           introPeriod: mockIntroPeriod,
           price: 1.99
-        )
+        ), 
+      entitlements: []
     )
     let productsById = [productId: product]
 
@@ -254,7 +267,7 @@ class PaywallLogicTests: XCTestCase {
     let productId = "id1"
     let products = [Product(
       name: "primary",
-      type: .appStore(.init(id: productId))
+      type: .appStore(.init(id: productId)), entitlements: []
     )]
     let mockIntroPeriod = MockIntroductoryPeriod(
       testSubscriptionPeriod: MockSubscriptionPeriod()
@@ -265,7 +278,8 @@ class PaywallLogicTests: XCTestCase {
           productIdentifier: productId,
           introPeriod: mockIntroPeriod,
           price: 1.99
-        )
+        ),
+      entitlements: []
     )
     let productsById = [productId: product]
 
@@ -294,7 +308,8 @@ class PaywallLogicTests: XCTestCase {
     let productId = "id1"
     let products = [Product(
       name: "primary",
-      type: .appStore(.init(id: productId))
+      type: .appStore(.init(id: productId)), 
+      entitlements: []
     )]
     let mockIntroPeriod = MockIntroductoryPeriod(
       testSubscriptionPeriod: MockSubscriptionPeriod()
@@ -305,7 +320,8 @@ class PaywallLogicTests: XCTestCase {
           productIdentifier: productId,
           introPeriod: mockIntroPeriod,
           price: 1.99
-        )
+        ), 
+      entitlements: []
     )
     let productsById = [productId: product]