+++ CHANGELOG.md
@@ -8,8 +8,11 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 
 - Removes `trigger_session_id` from `PaywallInfo` params.
 - `ProductInfo` is renamed to `Product` and the old `Product` class no longer exists.
+
+// TODO: Add more info here
 - Removes `Superwall.shared.subscriptionStatus` in favor of entitlements.
 - Changes `subscriptionStatus_didChange` to `activeEntitlements_didChange`.
+- Renames `productItems` to `products` in `PaywallInfo`.
 
 ## 3.7.4